backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
lxml==5.3.0
pillow==11.0.0
PyJWT==2.10.1
python-docx==1.1.2
reportlab==4.2.5
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise
drf-spectacular